practice/requirements.txt
fastapi
uvicorn[standard]
jinja2
tortoise-orm
aiosqlite
python-multipart
pandas
matplotlib
